+++ test/integration/verify/controls/google_cloud_scheduler_job.rb
@@ -14,8 +14,8 @@
 
 title 'Test GCP google_cloud_scheduler_job resource.'
 
-gcp_project_id = attribute(:gcp_project_id, default: 'gcp_project_id', description: 'The GCP project identifier.')
-scheduler_job = attribute('scheduler_job', default: {
+gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
+scheduler_job = input('scheduler_job', value: {
   "region": "us-central1",
   "name": "job-name",
   "description": "A description",