+++ test/integration/verify/controls/google_compute_global_operations.rb
@@ -14,8 +14,8 @@
 
 title 'Test GCP google_compute_global_operations resource.'
 
-gcp_project_id = attribute(:gcp_project_id, default: 'gcp_project_id', description: 'The GCP project identifier.')
-global_operation = attribute('global_operation', default: {
+gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
+global_operation = input('global_operation', value: {
   "name": "operation-1634799391539-5ced765030229-be5d5765-6623920f",
   "operationType": "compute.externalVpnGateways.delete"
 }, description: 'Global operation rule definition')