+++ test/integration/verify/controls/google_iam_organization_custom_roles.rb
@@ -14,9 +14,9 @@
 
 title 'Test GCP google_iam_organization_custom_roles resource.'
 
-gcp_organization_id = attribute(:gcp_organization_id, default: gcp_organization_id, description: 'The identifier of the organization that is the parent of the folder')
-gcp_organization_iam_custom_role_id = attribute(:gcp_organization_iam_custom_role_id, default: 'gcp_organization_iam_custom_role_id', description: 'The IAM custom role identifier.')
-gcp_enable_privileged_resources = attribute(:gcp_enable_privileged_resources, default:0, description:'Flag to enable privileged resources requiring elevated privileges in GCP.')
+gcp_organization_id = input(:gcp_organization_id, value: gcp_organization_id, description: 'The identifier of the organization that is the parent of the folder')
+gcp_organization_iam_custom_role_id = input(:gcp_organization_iam_custom_role_id, value: 'gcp_organization_iam_custom_role_id', description: 'The IAM custom role identifier.')
+gcp_enable_privileged_resources = input(:gcp_enable_privileged_resources, value:0, description:'Flag to enable privileged resources requiring elevated privileges in GCP.')
 control 'google_iam_organization_custom_roles-1.0' do
   impact 1.0
   title 'google_iam_organization_custom_roles resource test'