+++ test/integration/verify/controls/google_project_metrics.rb
@@ -14,8 +14,8 @@
 
 title 'Test GCP google_project_metrics resource.'
 
-gcp_project_id = attribute(:gcp_project_id, default: 'gcp_project_id', description: 'The GCP project identifier.')
-logging_metric = attribute('logging_metric', default: {
+gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
+logging_metric = input('logging_metric', default: {
   "name": "some/metric",
   "filter": "resource.type=gae_app AND severity>=ERROR",
   "metric_kind": "DELTA",