+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Change Log
 
-<!-- latest_release 1.10.27 -->
+<!-- latest_release 1.10.29 -->
+## [v1.10.29](https://github.com/inspec/inspec-gcp/tree/v1.10.29) (2022-06-16)
+
+#### Merged Pull Requests
+- input and value updated from controls [#407](https://github.com/inspec/inspec-gcp/pull/407) ([sa-progress](https://github.com/sa-progress))
+<!-- latest_release -->
+
+## [v1.10.28](https://github.com/inspec/inspec-gcp/tree/v1.10.28) (2022-06-15)
+
+#### Merged Pull Requests
+- RESOURCE-325-soc2-section-cc6-6-gcp [#411](https://github.com/inspec/inspec-gcp/pull/411) ([sa-progress](https://github.com/sa-progress))
+
 ## [v1.10.27](https://github.com/inspec/inspec-gcp/tree/v1.10.27) (2022-05-19)
 
 #### Merged Pull Requests
 - Do not raise an error when calling resource_id if @connection is nil [#416](https://github.com/inspec/inspec-gcp/pull/416) ([rbclark](https://github.com/rbclark))
-<!-- latest_release -->
 
 ## [v1.10.26](https://github.com/inspec/inspec-gcp/tree/v1.10.26) (2022-05-10)